Refactor SCRAM-SHA-1 unit test: separate generic vs SHA-1-specific tests
The goal of this change is to prepare the unit test structure for SHA-256 re-use.

A new abstract test class is created that now contains tests for:
- state machine
- malformed input
- negotiated properties
- channel binding behavior

Tests (and helpers) for protocol semantics, SASL contract behavior and algorithm-independent SCRAM behavior now live in that abstract class. SHA-1-specific cryptographic behavior, mechanism naming and success path details remain in the SCRAM-SHA-1 specific class.

The fake salt determinism, fake key generation and user enumeration protection responsibilities are deliberately not folded into the same abstract class, as that would cause us to end up with a "god test class". There is some duplication in those that is worth removing, but that will be the subject of a future commit.
